requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==3.0.2
Werkzeug==2.1.2
PyJWT==2.3.0
itsdangerous==2.1.2
cryptography==36.0.1
